In category repin does not shows errors in a separate page
In the category listing, when you click on the image then add to getlist, a pop-up window appears to re-pin the image. In that window, if there are errors, the errors are shown in the same window; before the errors where shown in a blanck page.
After that re-pin, the user goes to the profile.